refactor(arithmetic): Calculator -> ArithmeticCalculator로 이름 변경
