+++ src/main/java/calculator/ArithmeticCalculator.java
@@ -1,6 +1,6 @@
 package calculator;
 
-public class Calculator {
+public class ArithmeticCalculator {
 	int add(int num1, int num2) {
 		return num1 + num2;
 	}
